feat: Rename files and add dist files
