+++ src/Avatar/avatar.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
-import Avatar from "./Avatar";
+import Avatar from "./index";
 import Text from "../Typography/Text";
 
 export const names = [
