+++ src/Avatar/index.js
