+++ src/Typography/index.js
@@ -1,0 +1,11 @@
+import Text from "./Text";
+import Heading from "./Heading";
+import UnorderedList from "./UnorderedList";
+import ListItem from "./ListItem";
+
+export {
+    Text,
+    Heading,
+    UnorderedList,
+    ListItem
+}
